feat: Made code modular, invoked through client
